// src/FE/src/api/generated/models/inventoryScanResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Workslip.Api | v1
 * OpenAPI spec version: 1.0.0
 */
import type { InventoryBalanceResponse } from './inventoryBalanceResponse';

export interface InventoryScanResponse {
  materialId: string;
  name: string;
  sku: string;
  unit: string;
  balances: InventoryBalanceResponse[];
}